+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 ## Unreleased
+* Add aliases support.
+* Add default client constructors for Contentful and Github.
 
 ## v0.5.0
 ### Added